// Chapter_4/create_model_projeck_with_npx/Menu/Menu.tsx
import {AppContext} from "../context/app.context";
import {useContext} from "react";
import {FirstLevelMenuItem, PageItem} from "../interfaces/menu.interface";
import CoursesIcon from "./icons/courses.svg"
import {TopLevelCategory} from "../interfaces/page.interface";
import ServicesIcon from "./icons/servises.svg"
import ProductsIcon from "./icons/product.svg"
import BooksItem from "./icons/books.svg"
import styles from "./Menu.module.css"
import cn from "classnames";

const firstLevelMenu: FirstLevelMenuItem[] = [
    {route: 'courses', name: 'Курсы', icon: <CoursesIcon/>, id: TopLevelCategory.Courses},
    {route: 'services', name: 'Сервисы', icon: <ServicesIcon/>, id: TopLevelCategory.Services},
    {route: 'products', name: 'Продукты', icon: <ProductsIcon/>, id: TopLevelCategory.Products},
    {route: 'books', name: 'Книги', icon: <BooksItem/>, id: TopLevelCategory.Books},
];


export const Menu = (): JSX.Element => {
    const {menu, setMenu, firstCategory} = useContext(AppContext);

    const buildFirstLevel = () => {
        return (
            <>
                {firstLevelMenu.map(m => (
                    <div key={m.route}>
                        <a href={'/${m.route}'}>
                            <div className={cn(styles.firstLevel, {
                                [styles.firstLevelActive]: m.id == firstCategory
                            })}>
                                {m.icon}
                                <span>{m.name}</span>
                            </div>
                        </a>
                        {m.id == firstCategory && buildSecondLevel(m)}
                    </div>
                ))}
            </>
        )
    };
    const buildSecondLevel = (menuItem: FirstLevelMenuItem) => {
        return (
            <div className={styles.secondBlock}>
                {menu.map(m => (
                    <div key={m._id.secondCategory}>
                        <div className={styles.secondLevel}>
                            {m._id.secondCategory}
                        </div>
                        <div className={cn(styles.secondLevelBlock, {
                            [styles.secondLevelBlockOpened]: m.isOpened
                        })}>
                            {buildThirdLevel(m.pages, menuItem.route)}
                        </div>
                    </div>
                ))}
            </div>
        )
    };
    const buildThirdLevel = (pages: PageItem[], route: string) => {
        return (
            pages.map(p => (
                <a href={'/${route}/${p.alias}'} className={cn(styles.thirdLevel, {
                    [styles.thirdLevelActive]: false
                })}>
                    {p.category}
                </a>
            ))
        )
    };


    return (
        <div className={styles.menu}>
            {buildFirstLevel()}
        </div>
    )
}